+++ src/utilities.jl
@@ -1,5 +1,3 @@
-import Base: clamp, clamp!
-
 """
     gainloss(contemporary::SDMLayer{Bool}, future::SDMLayer{Bool})
 
@@ -19,15 +17,4 @@ function discretize(layer, n::Integer)
     map!(x -> x * (n - 1) + 1, categories.grid, categories.grid)
     map!(x -> isnan(x) ? zero(eltype(categories.grid)) : x, categories.grid)
     return convert(SDMLayer{Int}, categories)
-end
-
-function Base.clamp!(layer::SDMLayer{T}, lo::L, hi::H) where {T <: Number, L <: Number, H <: Number}
-    clamp!(layer.grid, lo, hi)
-    return layer
-end
-
-function Base.clamp(layer::SDMLayer{T}, lo::L, hi::H) where {T <: Number, L <: Number, H <: Number}
-    cp = copy(layer)
-    clamp!(cp, lo, hi)
-    return cp
 end